+++ components/consumable.py
@@ -9,7 +9,7 @@ from exceptions import Impossible
 from handlers.select_adjacent_handler import SelectAdjacentHandler
 from handlers.area_ranged_attack_handler import AreaRangedAttackHandler
 from handlers.single_ranged_attack_handler import SingleRangedAttackHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 import actions.item_action
 import color
 import components.ai.confused_ai
