--- handlers/input_handlers.py
@@ -1,40 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Optional, Type, Union
-
-import tcod
-
-from actions.action import Action
-
-
-ActionOrHandler = Union[Action, "BaseEventHandler"]
-"""An event handler return value which can trigger an action or switch active handlers.
-
-If a handler is returned then it will become the active handler for future events.
-If an action is returned it will be attempted and if it's valid then
-MainGameEventHandler will become the active handler.
-"""
-
-
-class BaseEventHandler(tcod.event.EventDispatch[ActionOrHandler]):
-    def handle_events(self, event: tcod.event.Event) -> BaseEventHandler:
-        """Handle an event and return the next active event handler."""
-        state = self.dispatch(event)
-        if isinstance(state, BaseEventHandler):
-            return state
-        assert not isinstance(state, Action), f"{self!r} can not handle actions."
-        return self
-
-    def on_render(self, console: tcod.console.Console) -> None:
-        raise NotImplementedError()
-
-    def ev_quit(self, event: tcod.event.Quit) -> Optional[Action]:
-        raise SystemExit()
-
-
-
-
-
-
-
-
