+++ handlers/inspect_item_handler.py
@@ -2,7 +2,7 @@ from engine import Engine
 from entity import Item
 from handlers.inventory_event_handler import InventoryEventHandler
 from handlers.event_handler import EventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
 
 import tcod
