+++ handlers/inventory_activate_handler.py
@@ -2,7 +2,7 @@ import color
 from actions.equip_action import EquipAction
 from entity import Item
 from handlers.inventory_event_handler import InventoryEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
