+++ handlers/inventory_binds_handler.py
@@ -1,7 +1,7 @@
 import color
 from entity import Item
 from handlers.inventory_event_handler import InventoryEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
