+++ handlers/inventory_drop_handler.py
@@ -1,7 +1,7 @@
 from actions.drop_item_action import DropItemAction
 from entity import Item
 from handlers.inventory_event_handler import InventoryEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
