+++ handlers/inventory_event_handler.py
@@ -3,7 +3,7 @@ import keys
 from engine import Engine
 from entity import Item
 from handlers.ask_user_event_handler import AskUserEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
 
 import tcod
