+++ handlers/level_up_event_handler.py
@@ -1,6 +1,6 @@
 import color
 from handlers.ask_user_event_handler import AskUserEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
