+++ handlers/select_adjacent_handler.py
@@ -3,7 +3,7 @@ from actions.action import Action
 from engine import Engine
 from handlers.ask_user_event_handler import AskUserEventHandler
 from typing import Optional
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 
 
 import tcod
