+++ handlers/select_index_handler.py
@@ -2,7 +2,7 @@ import color
 import keys
 from engine import Engine
 from handlers.ask_user_event_handler import AskUserEventHandler
-from handlers.input_handlers import ActionOrHandler
+from handlers.action_or_handler import ActionOrHandler
 from typing import Optional
